feat: added express server and ping route
