+++ package.json
@@ -13,14 +13,18 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/express": "^4.17.17",
     "@types/johnny-five": "^2.1.3",
+    "@types/morgan": "^1.9.4",
     "@types/node": "^20.4.2",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.1.6"
   },
   "dependencies": {
+    "express": "^4.18.2",
     "firmata": "^2.3.0",
     "johnny-five": "^2.1.0",
+    "morgan": "^1.10.0",
     "serialport": "^9.2.8"
   }
 }
