+++ src/lib/entities/state.entity.ts
