+++ src/routes/ping.route.ts
@@ -1,0 +1,7 @@
+import { Router } from "express";
+
+const router = Router();
+
+router.get("/ping", (_req, res) => res.json("pong"));
+
+export default router;
